gemini-github-integration/requirements.txt
# Requirements for Gemini GitHub Integration
# Core dependencies for Google Cloud Functions and GitHub integration

# Google Cloud and AI
google-cloud-aiplatform>=1.34.0
functions-framework>=3.4.0
vertexai>=1.34.0

# GitHub integration
PyGithub>=1.59.1
requests>=2.31.0

# Web framework
Flask>=2.3.3

# Data handling
pydantic>=2.4.2
python-dateutil>=2.8.2

# Logging and monitoring
google-cloud-logging>=3.8.0

# Security
cryptography>=41.0.4

# Development tools
pytest>=7.4.2
pytest-asyncio>=0.21.1
black>=23.9.1
flake8>=6.0.0
mypy>=1.5.1

# Optional: For local development
python-dotenv>=1.0.0
